feat: 홈 화면에 오늘의 구장 현황 표시
