+++ android/app/src/main/java/com/yagubogu/presentation/home/HomeViewModel.kt
@@ -9,6 +9,7 @@ import com.yagubogu.domain.model.Coordinate
 import com.yagubogu.domain.model.Distance
 import com.yagubogu.domain.model.Stadium
 import com.yagubogu.domain.model.Stadiums
+import com.yagubogu.domain.model.Team
 import com.yagubogu.domain.repository.CheckInsRepository
 import com.yagubogu.domain.repository.LocationRepository
 import com.yagubogu.domain.repository.MemberRepository
@@ -16,6 +17,10 @@ import com.yagubogu.domain.repository.StadiumRepository
 import com.yagubogu.domain.repository.StatsRepository
 import com.yagubogu.presentation.home.model.CheckInUiEvent
 import com.yagubogu.presentation.home.model.HomeUiModel
+import com.yagubogu.presentation.home.model.StadiumStatsUiModel
+import com.yagubogu.presentation.home.model.TeamOccupancyRate
+import com.yagubogu.presentation.home.model.TeamOccupancyRates
+import com.yagubogu.presentation.home.model.TeamOccupancyStatus
 import com.yagubogu.presentation.util.livedata.MutableSingleLiveData
 import com.yagubogu.presentation.util.livedata.SingleLiveData
 import kotlinx.coroutines.Deferred
@@ -37,8 +42,21 @@ class HomeViewModel(
     private val _checkInUiEvent = MutableSingleLiveData<CheckInUiEvent>()
     val checkInUiEvent: SingleLiveData<CheckInUiEvent> get() = _checkInUiEvent
 
+    private val _stadiumStatsUiModel: MutableLiveData<StadiumStatsUiModel> = MutableLiveData()
+    val stadiumStatsUiModel: LiveData<StadiumStatsUiModel> get() = _stadiumStatsUiModel
+
     init {
         fetchMemberInformation(MEMBER_ID, YEAR)
+
+        _stadiumStatsUiModel.value =
+            StadiumStatsUiModel(
+                "로딩중",
+                listOf(TeamOccupancyStatus(Team.LG, 0.0)),
+            )
+
+//        val today = LocalDate.now()
+        val today = LocalDate.of(2025, 7, 25)
+        fetchStadiumStats(DUMMY_STADIUM_ID, today)
     }
 
     fun checkIn() {
@@ -91,6 +109,59 @@ class HomeViewModel(
         }
     }
 
+    private fun fetchStadiumStats(
+        stadiumId: Long,
+        date: LocalDate,
+    ) {
+        viewModelScope.launch {
+            val teamOccupancyRatesResult: Result<TeamOccupancyRates> =
+                statsRepository.getTeamOccupancyRates(stadiumId, date)
+            teamOccupancyRatesResult
+                .onSuccess { teamOccupancyRates: TeamOccupancyRates ->
+                    val teamOccupancyStatuses: List<TeamOccupancyStatus> =
+                        teamOccupancyRates.rates.map { teamOccupancyRate: TeamOccupancyRate ->
+                            val team: Team = Team.getById(teamOccupancyRate.teamId)
+                            TeamOccupancyStatus(
+                                team,
+                                teamOccupancyRate.occupancyRate,
+                            )
+                        }
+
+                    val refinedTeamStatuses: List<TeamOccupancyStatus> =
+                        refineTeamStatus(teamOccupancyStatuses)
+                    _stadiumStatsUiModel.value =
+                        StadiumStatsUiModel(
+                            teamOccupancyRates.stadiumName,
+                            refinedTeamStatuses,
+                        )
+                }.onFailure { exception: Throwable ->
+                    Log.e(TAG, "API 호출 실패", exception)
+                }
+        }
+    }
+
+    private fun refineTeamStatus(teamStatuses: List<TeamOccupancyStatus>): List<TeamOccupancyStatus> =
+        if (teamStatuses.isEmpty()) {
+            teamStatuses
+        } else {
+            when {
+                teamStatuses.size <= MAX_LEGEND_TEAM_SIZE -> teamStatuses
+                else -> {
+                    val topLegendTeamStatues: List<TeamOccupancyStatus> =
+                        teamStatuses.take(MAX_LEGEND_TEAM_SIZE)
+                    val etcPercentage: Double =
+                        FULL_PERCENTAGE - topLegendTeamStatues.sumOf { it.percentage }
+
+                    val etcTeamStatus =
+                        TeamOccupancyStatus(
+                            team = null,
+                            percentage = etcPercentage,
+                        )
+                    topLegendTeamStatues + etcTeamStatus
+                }
+            }
+        }
+
     private fun handleCheckIn(currentCoordinate: Coordinate) {
         viewModelScope.launch {
             val stadiumsResult: Result<Stadiums> = stadiumRepository.getStadiums()
@@ -134,5 +205,9 @@ class HomeViewModel(
         private const val THRESHOLD_IN_METERS = 2200.0 // TODO: 300.0 으로 변경
         private const val MEMBER_ID = 5009L
         private const val YEAR = 2025
+
+        private const val DUMMY_STADIUM_ID = 2L // 잠실구장
+        private const val MAX_LEGEND_TEAM_SIZE = 2
+        private const val FULL_PERCENTAGE = 100
     }
 }
