+++ android/app/src/main/res/values/strings.xml
@@ -24,6 +24,8 @@
     <string name="home_location_fetch_failed_message">위치를 불러오는 데 실패했습니다.</string>
     <string name="home_check_in_success_message">직관 인증이 완료되었습니다!</string>
     <string name="home_check_in_failure_message">구장 근처에서만 인증할 수 있어요.</string>
+    <string name="home_stadium_pie_chart_title">%s 현황</string>
+    <string name="home_stadium_refresh_time">%02d:%02d 기준</string>
 
     <string name="stats_tab_my_stats">내 통계</string>
     <string name="stats_tab_stadium_stats">오늘의 구장 통계</string>
@@ -40,7 +42,4 @@
     <string name="stats_my_team_emoji">🏆</string>
     <string name="stats_my_lucky_stadium_emoji">🍀</string>
     <string name="stats_my_no_lucky_stadium">없어요..</string>
-
-    <string name="stats_stadium_pie_chart_title">%s 현황</string>
-    <string name="stats_stadium_refresh_time">%02d:%02d 기준</string>
 </resources>
